add absolute and differential encoder
