+++ march_hardware/src/Encoder.cpp
@@ -7,29 +7,12 @@
 
 namespace march
 {
-Encoder::Encoder(int numberOfBits, int minPositionIU, int maxPositionIU, int zeroPositionIU, float safetyMarginRad)
+Encoder::Encoder(int numberOfBits)
 {
   ROS_ASSERT_MSG(numberOfBits > 0 && numberOfBits <= 32, "Encoder resolution of %d is not within range (0, 32)",
                  numberOfBits);
   this->totalPositions = static_cast<int>(pow(2, numberOfBits) - 1);
-
-  ROS_ASSERT_MSG(safetyMarginRad >= 0, "SafetyMarginRad %f is below zero", safetyMarginRad);
-
-  this->safetyMarginRad = safetyMarginRad;
   this->slaveIndex = -1;
-  this->upperHardLimitIU = maxPositionIU;
-  this->lowerHardLimitIU = minPositionIU;
-  this->zeroPositionIU = zeroPositionIU;
-  int safetyMarginIU = RadtoIU(safetyMarginRad) - this->zeroPositionIU;
-  this->upperSoftLimitIU = this->upperHardLimitIU - safetyMarginIU;
-  this->lowerSoftLimitIU = this->lowerHardLimitIU + safetyMarginIU;
-
-  ROS_ASSERT_MSG(this->lowerSoftLimitIU < this->upperSoftLimitIU,
-                 "Invalid range of motion. Safety margin too large or "
-                 "min/max position invalid. lowerSoftLimit: %i IU, upperSoftLimit: "
-                 "%i IU. lowerHardLimit: %i IU, upperHardLimit %i IU. safetyMargin: %f rad = %i IU",
-                 this->lowerSoftLimitIU, this->upperSoftLimitIU, this->lowerHardLimitIU, this->upperHardLimitIU,
-                 this->safetyMarginRad, safetyMarginIU);
 }
 
 float Encoder::getAngleRad(uint8_t ActualPositionByteOffset)
@@ -67,55 +50,4 @@ int Encoder::getSlaveIndex() const
 {
   return this->slaveIndex;
 }
-
-bool Encoder::isWithinHardLimitsIU(int positionIU)
-{
-  return (positionIU > this->lowerHardLimitIU && positionIU < this->upperHardLimitIU);
-}
-
-bool Encoder::isWithinSoftLimitsIU(int positionIU)
-{
-  return (positionIU > this->lowerSoftLimitIU && positionIU < this->upperSoftLimitIU);
-}
-
-bool Encoder::isValidTargetIU(int currentIU, int targetIU)
-{
-  if (this->isWithinSoftLimitsIU(targetIU))
-  {
-    return true;
-  }
-
-  if (currentIU >= this->getUpperSoftLimitIU())
-  {
-    return (targetIU <= currentIU) && (targetIU > this->getLowerSoftLimitIU());
-  }
-
-  if (currentIU <= this->getLowerSoftLimitIU())
-  {
-    return (targetIU >= currentIU) && (targetIU < this->getUpperSoftLimitIU());
-  }
-
-  return false;
-}
-
-int Encoder::getUpperSoftLimitIU() const
-{
-  return upperSoftLimitIU;
-}
-
-int Encoder::getLowerSoftLimitIU() const
-{
-  return lowerSoftLimitIU;
-}
-
-int Encoder::getUpperHardLimitIU() const
-{
-  return upperHardLimitIU;
-}
-
-int Encoder::getLowerHardLimitIU() const
-{
-  return lowerHardLimitIU;
-}
-
 }  // namespace march
